+++ docs/changelog.md
@@ -14,6 +14,10 @@ Versioning currently follows `X.Y.Z` where
 
 - Entities can be compared and sorted by ID now.
 
+### Changed
+
+- (internal) migrate to src layout
+
 ## \[1.13.7\] - 2024-09-17
 
 ### Fixed