+++ docs/changelog.md
@@ -13,6 +13,7 @@ Versioning currently follows `X.Y.Z` where
 ### Added
 
 - Entities can be compared and sorted by ID now.
+- Caching for list_not_existing_storage_directories.
 - (experimental) add initial code for a resource based application dispatch
 - (experimental) new app_runner cli that integrates all commands into a single interface